requirements.txt
fastapi==0.128.0
uvicorn==0.27.0
pydantic==2.12.5
pydantic-settings==2.1.0
pydantic_core==2.41.5
SQLAlchemy==2.0.45
aiosqlite==0.19.0
python-jose==3.3.0
bcrypt==4.0.1
passlib==1.7.4
python-multipart==0.0.6
Jinja2==3.1.3
pytest
httpx==0.26.0
pytest-asyncio
python-dotenv==1.2.1
email-validator==2.3.0
dotenv==0.9.9